now connected to pinecone, images are filtered then uploaded to pinecone

+++ src/main/java/com/example/demo/entities/ImagesEmbedding.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @Data
 public class ImagesEmbedding {
 
-    ArrayList<Double> embedding;
+    ArrayList<Float> embedding;
     
 }
